backend/requirements.txt
fastapi
uvicorn
gigachat
sqlalchemy
pydantic
python-dotenv
pymupdf